fix(client): split the field-notes file, and stop the drawer stealing focus back
CI's structure gate measures new files against develop, where
GardenDetailFieldNotes.tsx is new and was 406 lines against a 350 limit. My
local runs compared against the previous commit, so they saw it as modified and
let it through — that one violation is what failed all six red jobs. The record
drawer and its media mosaic move to GardenDetailNoteRecord.tsx; the section and
its tile grid stay.

The drawer focused its close control from an inline callback ref, which is a new
function on every render, so React re-ran it every time. Closing the image
viewer re-renders the drawer, which pulled focus out of the photo the reader had
just been looking at and back to Close. It now focuses from an effect keyed on
`open`, and a test covers both halves: focus lands on open, and stays put across
a re-render from inside or above. The test fails against the old ref.
